[mz] stub out an error getting field labels for an invalid field
For some reason metabase is fetching fields at index 15-18 for the foreign-keys
sync, when the column that is being returned only has fields 13-15 set.

I'm not sure why this is happening, but since we don't have foreign keys anyway
I'm going to just stub this out with some field names that look pretty obvious
in the hope that we can debug it if/when it comes back up.